Provide better progress reporting to the user
Fetching each workflow run requires an HTTP per workflow and this can
become quite slow, even for only a few days of data. Provide the user
with a progress bar, with an ETA via tty/progressbar so that they can
see progress.

Also, this helped me notice and fix that we were never fetching the last
page of data.